Improved Readability + Simplified Grid System
This morning I've been cleaning up and commenting more sections of my blueprints.

I took this opportunity to rework how some of the grid is setup because it was making implementing neighbours needlessly messy. Prior to today I've been using a small number of actors to store different aspects of the grid to help organize components.

However the main goal of my work is to have a grid singleton that the game can hook into at will. This means now I have a better idea of the grid's final end state, I need to streamline grid systems back into a singleton. This is not fully completed just yet but I am most of the way there now.